+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [16.3.0] - 2026-02-11
+
+### Changed
+
+- Add configuration property "jeap.processcontext.kafka.message-consumer-paused" (default false) to be able to pause
+  processing of messages in preparation of next major version
+- Remove property jeap.processcontext.process-updates.auto-start
+
 ## [16.2.1] - 2026-01-28
 
 ### Changed